:label: Rename: 공기질 데이터 저장 컨트롤러, 서비스 패키지 안으로 이동
